+++ pkgs/checks/lib/src/extensions/function.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:checks/context.dart';
+import '../../context.dart';
 
 extension FunctionChecks<T> on Subject<T Function()> {
   /// Expects that a function throws synchronously when it is called.