+++ pkgs/test/lib/src/executable.dart
@@ -9,7 +9,6 @@ import 'package:test_core/src/runner/hack_register_platform.dart'; // ignore: im
 
 import 'runner/browser/platform.dart';
 import 'runner/node/platform.dart';
-import 'runner/wasm/platform.dart';
 
 Future<void> main(List<String> args) async {
   registerPlatformPlugin([Runtime.nodeJS], NodePlatform.new);
@@ -19,9 +18,6 @@ Future<void> main(List<String> args) async {
     Runtime.firefox,
     Runtime.safari,
   ], BrowserPlatform.start);
-  registerPlatformPlugin([
-    Runtime.experimentalChromeWasm,
-  ], BrowserWasmPlatform.start);
 
   await executable.main(args);
 }