+++ pkgs/test/tool/host.dart
@@ -9,7 +9,6 @@ import 'dart:async';
 import 'dart:convert';
 
 import 'package:js/js.dart';
-import 'package:js/js_util.dart' as js_util;
 import 'package:stack_trace/stack_trace.dart';
 import 'package:stream_channel/stream_channel.dart';
 import 'package:test/src/runner/browser/dom.dart' as dom;
@@ -138,7 +137,7 @@ void main() {
 
     // Send periodic pings to the test runner so it can know when the browser is
     // paused for debugging.
-    Timer.periodic(Duration(seconds: 1),
+    Timer.periodic(const Duration(seconds: 1),
         (_) => serverChannel.sink.add({'command': 'ping'}));
 
     var play = dom.document.querySelector('#play');
@@ -168,7 +167,7 @@ MultiChannel<dynamic> _connectToServer() {
   var webSocket =
       dom.createWebSocket(_currentUrl.queryParameters['managerUrl']!);
 
-  var controller = StreamChannelController(sync: true);
+  var controller = StreamChannelController<Object?>(sync: true);
   webSocket.addEventListener('message', allowInterop((message) {
     controller.local.sink
         .add(jsonDecode((message as dom.MessageEvent).data as String));
@@ -206,7 +205,7 @@ StreamChannel<dynamic> _connectToIframe(String url, int id) {
   dom.window.console.log('Starting suite $suiteUrl');
   var iframe = dom.createHTMLIFrameElement();
   _iframes[id] = iframe;
-  var controller = StreamChannelController(sync: true);
+  var controller = StreamChannelController<Object?>(sync: true);
 
   late dom.Subscription windowSubscription;
   windowSubscription =
@@ -219,9 +218,7 @@ StreamChannel<dynamic> _connectToIframe(String url, int id) {
     if (message.origin != dom.window.location.origin) return;
     // Disambiguate between frames for different test suites.
     // Depending on the source type, the `location.href` may be missing.
-    var location = js_util.getProperty(message.source, 'location') as Object?;
-    if (location == null) return;
-    if (js_util.getProperty(location, 'href') != iframe.src) return;
+    if (message.source.location?.href != iframe.src) return;
 
     message.stopPropagation();
     windowSubscription.cancel();